refactor: create new UserApiKeysTable component
